ember-power-selected selected option test
- The documentation [says](https://github.com/MichalBryxi/ember-bootstrap-power-select#advanced-usage):

>The power-select's selected, disabled and placeholder properties and the onchange action are already wired up to the controlling form.element for you

I tried, it did not work. So I tried to poke deeper and found these:
 - https://github.com/kaliber5/ember-bootstrap-power-select/blob/6566374b18ce372a255b87f2fac76cebc28f0f06/addon/templates/components/bs-form/element/control/power-select.hbs#L4
 - https://github.com/kaliber5/ember-bootstrap-power-select/blob/6566374b18ce372a255b87f2fac76cebc28f0f06/addon/templates/components/bs-form/element/control/power-select.hbs#L31

So I tried to alter my code to pass `value` property and it worked:

```
{{#bs-form model=this as |form|}}
    {{form.element controlType="power-select" options=options value=selected}}
  {{/bs-form}}
```

So I'm proposing this failing test to alter the behaviour to be in line with official [ember-power-select docs](http://www.ember-power-select.com/docs/api-reference). And I think that property should be changed to from `value` to `selected`.